Added Rails compatibility
As an alternative to "Base.establish_connection", added
use_existing_connection to Agent and User classes.
The method takes an established connection as an argument.
With this, it becomes possible to add a database through the
database.yml file in a Rails app, and pass that connection to QUI.
For example, you could define queuemetrics: in the yml file,
and call Qui::Agent.use_existing_connection(:queuemetrics).
Doing this allows you to maintain the connection to the Rails
database while you interact with QUI.